Initial commit.
Well kinda... I rm'd the initial directory before pushing and the .git directory was not recoverable. Some of the history is lost, but it was really a run up to this point.

There are three individual C shared object libs that get created. The libs each have a Python module that provides interoperability using ctypes. Notebooks illustrate usage of each.